BREAKING CHANGE: upgrade react18 template to react 18 props
The react template now allows users to clear the chain by passing `true`
to the `toString()` fn. This removes all but the first item in the
array.

Added a `clear()` fn to the react template, which clears every element
in the chain array.

Cleaned up a few types that were too explicit to make rome linter happy.

BREAKING CHANGE: Added all the common html elements that are part of
the react template to the list of selectors that will not be generated
so it's much less likely to create a duplicate class member if the
input css has a selector called `.nav` or something.

Cleaned up a few things in package.json.